requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database (100% Async)
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0

# Validation and Settings
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# System monitoring
psutil==5.9.6

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
